+++ packages/text-selection-toolbar/web/src/index.js
@@ -1,3 +1,5 @@
 import TextSelectionListener from './TextSelectionListener.jsx';
 import ViewerInlineToolBar from './ViewerInlineToolBar.jsx';
-export { TextSelectionListener, ViewerInlineToolBar };
+import TwitterButton from './TwitterButton.jsx';
+
+export { TextSelectionListener, ViewerInlineToolBar, TwitterButton };